+++ GeoChat/package.json
@@ -13,6 +13,7 @@
     "react": "16.5.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
     "react-native-dotenv": "^0.2.0",
+    "react-native-elements": "^1.1.0",
     "react-native-gesture-handler": "^1.3.0",
     "react-navigation": "^3.11.0",
     "react-redux": "6.0.1",
