+++ GeoChat/screens/Chatroom.js
@@ -2,6 +2,7 @@ import React from 'react';
 import {View, Text, Button, TextInput, StyleSheet} from 'react-native';
 import * as firebase from 'firebase';
 import {Link} from 'react-router-native';
+import styles from '../styles/Chatroom'
 
 export default class Chatroom extends React.Component{
 
@@ -65,17 +66,17 @@ export default class Chatroom extends React.Component{
 
 }
 
-const styles = StyleSheet.create({
-    container : {
-        flex:1,
-        marginTop : 30,
-    },
+// const styles = StyleSheet.create({
+//     container : {
+//         flex:1,
+//         marginTop : 30,
+//     },
 
-    input: {
-        position: 'relative',
-        fontSize: 15,
-        backgroundColor: 'lightblue',
-        padding: 3,
-        borderRadius: 5,
-    }
-})
+//     input: {
+//         position: 'relative',
+//         fontSize: 15,
+//         backgroundColor: 'lightblue',
+//         padding: 3,
+//         borderRadius: 5,
+//     }
+// })
