+++ GeoChat/screens/LoadingScreen.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import {View, Text, Image, ActivityIndicator, StyleSheet} from 'react-native';
 import TempLogo from '../assets/TempLogo.png'
+import styles from '../styles/LoadingScreen'
 
 export default function LoadingScreen(){
     return(
@@ -12,10 +13,3 @@ export default function LoadingScreen(){
     );
 }
 
-const styles = StyleSheet.create({
-    container : {
-        height:"100%",
-        width : "100%",
-        backgroundColor : "#00B3DA"
-    }
-})
