+++ GeoChat/styles/Chatroom.js
@@ -1,0 +1,22 @@
+
+import { StyleSheet } from 'react-native';
+const styles = StyleSheet.create({
+
+    container : {
+        flex:1,
+        marginTop : 30,
+    },
+
+    input: {
+        position: 'relative',
+        fontSize: 15,
+        backgroundColor: 'lightblue',
+        padding: 3,
+        borderRadius: 5,
+    }
+
+
+});
+
+
+export default styles;
