+++ GeoChat/styles/LoadingScreen.js
@@ -1,0 +1,12 @@
+
+import { StyleSheet } from 'react-native';
+const styles = StyleSheet.create({
+    container : {
+        height:"100%",
+        width : "100%",
+        backgroundColor : "#00B3DA"
+    }
+});
+
+
+export default styles;
