+++ GeoChat/styles/Settings.js
@@ -1,0 +1,9 @@
+
+import { StyleSheet } from 'react-native';
+const styles = StyleSheet.create({
+
+
+});
+
+
+export default styles;
