+++ GeoChat/styles/createChatroom.js
@@ -1,0 +1,23 @@
+
+import { StyleSheet } from 'react-native';
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+    },
+  textBox: {
+    borderLeftWidth: 3,
+    borderTopWidth: 3,
+    borderRightWidth: 3,
+    borderBottomWidth: 3,
+    width: "95%",
+    height: 100,
+  }
+
+
+});
+
+
+export default styles;
